remove more sensitive data

+++ scripts/profilizer.sh
@@ -11,7 +11,8 @@ rm -rf tmp_profile/domain_to_categories.sqlite
 rm -rf tmp_profile/domain_to_categories.sqlite-journal
 rm -rf tmp_profile/favicons.sqlite-shm
 rm -rf tmp_profile/features
-rm -rf tmp_profile/logins.db
+rm -rf tmp_profile/logins.*
+rm -rf tmp_profile/key*
 rm -rf tmp_profile/places_semantic.sqlite
 rm -rf tmp_profile/places_semantic.sqlite-wal
 rm -rf tmp_profile/places.sqlite-shm
@@ -24,5 +25,9 @@ rm -rf tmp_profile/tabnotes.sqlite
 rm -rf tmp_profile/tabnotes.sqlite-wal
 rm -rf tmp_profile/targeting.snapshot.json
 rm -rf tmp_profile/weave
+if [[ -e "../profile.zip" ]]; then
+  echo "Error: profile.zip already exists. Remove it first." >&2
+  exit 1
+fi
 (cd tmp_profile && zip -9r ../profile.zip .)
 rm -rf tmp_profile
